split_main_on_modules
